refactor: 重构文件列表视图并移除无关代码
