+++ catfish/settings.py
@@ -138,7 +138,7 @@ LOGGING = {
     'disable_existing_loggers': False,
     'formatters': {
         'verbose': {
-            'format': '[%(asctime)s] [%(levelname)s] [%(name)s] >>> %(message)s'
+            'format': '[%(asctime)s] [%(levelname)s] [%(name)s] %(message)s'
         },
         'simple': {
             'format': '%(levelname)s %(message)s'
